Fix errors thrown when the portal in AlignedTrigger is null
This could be a race condition or something else. Either way, the
storePortal ref callback is called with a null portal when the
portal is unmounted. If for some reason, the AlignedTrigger still
tries to open after that, it must not install the 'mousemove'
event handler. It did this before and then failed to install the
target rect properly because this.portal.node did not exist.

This commit adds a target/portal check before doing this. It also
ensures that event handlers are added as late and removed as early
as possible to reduce the likeliness of other errors leaving
event handlers open.